Time: 0 ms (100.00%) | Memory: 6.2 MB (76.88%) - LeetSync

+++ 374-guess-number-higher-or-lower/guess-number-higher-or-lower.cpp
@@ -10,8 +10,8 @@
 class Solution {
 public:
     int guessNumber(int n) {
-        int left = 1, right = n;
-        int mid;
+        long left = 1, right = n;
+        long mid;
         while (left<=right){
             mid = left + (right-left)/2;
             if (guess(mid)==0){
